# offchain/backend-go/docs/swagger.yaml
basePath: /v1
definitions:
  github_com_vndc_backend_internal_models.ApproveKYCRequest:
    properties:
      level:
        description: Level is 1 (basic), 2 (standard), or 3 (advanced).
        example: 2
        type: integer
    type: object
  github_com_vndc_backend_internal_models.AssignRoleRequest:
    properties:
      role:
        description: 'Role must be one of: USER, MODERATOR, ADMIN.'
        example: MODERATOR
        type: string
    type: object
...
  github_com_vndc_backend_internal_models.ErrorBody:
    properties:
      code:
        example: VALIDATION_ERROR
        type: string
      details: {}
      message:
        example: wallet query param required
        type: string
    type: object
  github_com_vndc_backend_internal_models.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorBody'
      meta:
        $ref: '#/definitions/github_com_vndc_backend_internal_models.ResponseMeta'
      success:
        example: false
        type: boolean
    type: object
  github_com_vndc_backend_internal_models.KYCDocumentInfo:
    properties:
      document_ref:
        example: enc:s3::bucket/usr123/passport.enc
        type: string
      review_note:
        example: Document blurry — please resubmit
        type: string
      reviewed_at:
        type: string
      submitted_at:
        example: "2026-04-24T07:00:00Z"
        type: string
      type:
        example: PASSPORT
        type: string
    type: object
...
  github_com_vndc_backend_internal_models.MessageResponse:
    properties:
      message:
        example: Operation completed successfully
        type: string
      success:
        example: true
        type: boolean
    type: object
  github_com_vndc_backend_internal_models.PaginationMeta:
    properties:
      has_next:
        example: true
        type: boolean
      has_prev:
        example: false
        type: boolean
      page:
        example: 1
        type: integer
      page_size:
        example: 20
        type: integer
      pages:
        example: 5
        type: integer
      total:
        example: 100
        type: integer
    type: object
...
  github_com_vndc_backend_internal_models.ResponseMeta:
    properties:
      request_id:
        example: 01HWXYZ1234567890ABCDEF
        type: string
      timestamp:
        example: "2026-04-24T07:00:00Z"
        type: string
    type: object
...
  github_com_vndc_backend_internal_models.SuspendUserRequest:
    properties:
      reason:
        description: Reason must describe why the account is being suspended (10–500
          characters).
        example: Multiple violations of community guidelines detected
        type: string
    type: object
...
  github_com_vndc_backend_internal_models.UserInfo:
    properties:
      avatar_uri:
        example: https://cdn.example.com/avatars/alice.png
        type: string
      bio:
        example: Blockchain enthusiast
        type: string
      country:
        example: VN
        type: string
      created_at:
        example: "2026-01-01T00:00:00Z"
        type: string
      email:
        example: alice@example.com
        type: string
      email_verified:
        example: true
        type: boolean
      full_name:
        example: Alice Nguyen
        type: string
      id:
        example: 65f1a2b3c4d5e6f7a8b9c0d1
        type: string
      kyc_documents:
        items:
          $ref: '#/definitions/github_com_vndc_backend_internal_models.KYCDocumentInfo'
        type: array
      kyc_level:
        example: 2
        type: integer
      kyc_status:
        example: VERIFIED
        type: string
      language:
        example: vi
        type: string
      last_login_at:
        type: string
      login_count:
        example: 42
        type: integer
      metadata:
        additionalProperties: {}
        type: object
      referral_code:
        example: ALICE2026
        type: string
      roles:
        example:
        - '["USER"]'
        items:
          type: string
        type: array
      status:
        example: ACTIVE
        type: string
      timezone:
        example: Asia/Ho_Chi_Minh
        type: string
      two_factor_enabled:
        example: false
        type: boolean
      updated_at:
        example: "2026-04-24T07:00:00Z"
        type: string
      username:
        example: alice
        type: string
      wallet_address:
        example: 0xd8dA6BF26964aF9D7eEd9e03E53415D37aA96045
        type: string
    type: object
  github_com_vndc_backend_internal_models.UserListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/github_com_vndc_backend_internal_models.UserInfo'
        type: array
      meta:
        $ref: '#/definitions/github_com_vndc_backend_internal_models.ResponseMeta'
      pagination:
        $ref: '#/definitions/github_com_vndc_backend_internal_models.PaginationMeta'
      success:
        example: true
        type: boolean
    type: object
  github_com_vndc_backend_internal_models.UserResponse:
    properties:
      data:
        $ref: '#/definitions/github_com_vndc_backend_internal_models.UserInfo'
      meta:
        $ref: '#/definitions/github_com_vndc_backend_internal_models.ResponseMeta'
      success:
        example: true
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: dev@vndc.io
    name: VNDC Platform Team
  description: |-
    Off-chain backend for the VNDC token platform.

    ## Authentication
    The API uses **wallet-based authentication** (Sign-In With Ethereum / EIP-191).
    All protected endpoints require an `Authorization: Bearer <access_token>` header.

    ### Login flow
    1. `GET /v1/auth/challenge?wallet=0x…`  — obtain a nonce-embedded message
    2. Sign the message with your wallet (personal_sign / EIP-191)
    3. `POST /v1/auth/login` — submit the signature; receive a JWT pair
    4. *(if 2FA is enabled)* `POST /v1/auth/2fa/complete` — verify TOTP code
    5. Pass the `access_token` as `Authorization: Bearer <token>` on every request
    6. `POST /v1/auth/refresh` — rotate tokens before access token expires (15 min)

    ## Amount encoding
    All token amounts are expressed in **wei** as **decimal strings** — never floats.
    Example: `"1000000000000000000"` = 1 VNDC (18 decimals).
  license:
    name: Proprietary
    url: https://vndc.io/terms
  title: VNDC Backend API
  version: 1.0.0
paths:
...
  /users:
    get:
      description: |-
        Returns a paginated, searchable list of all users on the platform.
        Filter by `status`, `kyc_status`, or `role`. Search by wallet, email, or username.
      parameters:
      - description: Filter by account status
        enum:
        - ACTIVE
        - SUSPENDED
        - BANNED
        - PENDING_VERIFICATION
        in: query
        name: status
        type: string
      - description: Filter by KYC status
        enum:
        - NONE
        - PENDING
        - VERIFIED
        - REJECTED
        in: query
        name: kyc_status
        type: string
      - description: Filter by role
        enum:
        - USER
        - MODERATOR
        - ADMIN
        in: query
        name: role
        type: string
      - description: Full-text search (wallet / email / username)
        in: query
        name: search
        type: string
      - description: 'Page number (default: 1)'
        example: 1
        in: query
        name: page
        type: integer
      - description: 'Items per page (max: 100)'
        example: 20
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.UserListResponse'
        "401":
          description: Missing or invalid Bearer token
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
        "403":
          description: Requires ADMIN role
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List users (admin)
      tags:
      - Users
  /users/{id}:
    get:
      description: |-
        Returns the full profile for any user on the platform, including sensitive fields
        such as KYC documents, suspension reason, and login statistics.
      parameters:
      - description: MongoDB ObjectID or UUID
        example: 65f1a2b3c4d5e6f7a8b9c0d1
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.UserResponse'
        "401":
          description: Missing or invalid Bearer token
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
        "403":
          description: Requires ADMIN role
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user by ID (admin)
      tags:
      - Users
  /users/{id}/kyc/approve:
    post:
      consumes:
      - application/json
      description: |-
        Marks the target user's KYC as `VERIFIED` at the specified level (1–3).
        | Level | Description |
        |-------|-------------|
        | 1     | Basic — email + phone confirmed |
        | 2     | Standard — government-issued ID |
        | 3     | Advanced — face-match + liveness check |
      parameters:
      - description: Target user ID
        in: path
        name: id
        required: true
        type: string
      - description: KYC level to grant (1–3)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_vndc_backend_internal_models.ApproveKYCRequest'
      produces:
      - application/json
      responses:
        "200":
          description: KYC approved
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.MessageResponse'
        "400":
          description: Invalid level
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
        "401":
          description: Missing or invalid Bearer token
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
        "403":
          description: Requires ADMIN role
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Approve KYC verification (admin)
      tags:
      - Users
  /users/{id}/roles:
    post:
      consumes:
      - application/json
      description: |-
        Adds the specified RBAC role to the target user's role list.
        Has no effect if the user already holds that role.
      parameters:
      - description: Target user ID
        in: path
        name: id
        required: true
        type: string
      - description: 'Role to assign: USER | MODERATOR | ADMIN'
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_vndc_backend_internal_models.AssignRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Role assigned
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.MessageResponse'
        "400":
          description: Invalid role value
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
        "401":
          description: Missing or invalid Bearer token
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
        "403":
          description: Requires ADMIN role
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
        "409":
          description: User already has this role
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Assign a role to a user (admin)
      tags:
      - Users
  /users/{id}/roles/{role}:
    delete:
      description: Removes the specified RBAC role from the target user. The base
        `USER` role cannot be removed.
      parameters:
      - description: Target user ID
        in: path
        name: id
        required: true
        type: string
      - description: Role to remove
        enum:
        - MODERATOR
        - ADMIN
        in: path
        name: role
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Role removed
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.MessageResponse'
        "400":
          description: Cannot remove base USER role
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
        "401":
          description: Missing or invalid Bearer token
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
        "403":
          description: Requires ADMIN role
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
        "404":
          description: User not found or does not have this role
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Remove a role from a user (admin)
      tags:
      - Users
  /users/{id}/suspend:
    post:
      consumes:
      - application/json
      description: |-
        Sets account status to `SUSPENDED` and immediately revokes all active sessions.
        The user will be unable to log in until the account is unsuspended.
        Cannot suspend your own account.
      parameters:
      - description: Target user ID
        in: path
        name: id
        required: true
        type: string
      - description: Mandatory suspension reason (10–500 chars)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_vndc_backend_internal_models.SuspendUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Account suspended
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.MessageResponse'
        "400":
          description: Cannot suspend own account / invalid reason
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
        "401":
          description: Missing or invalid Bearer token
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
        "403":
          description: Requires ADMIN role
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Suspend a user account (admin)
      tags:
      - Users
  /users/{id}/unsuspend:
    post:
      description: Restores account status to `ACTIVE`. The user can log in again
        immediately.
      parameters:
      - description: Target user ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Account unsuspended
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.MessageResponse'
        "401":
          description: Missing or invalid Bearer token
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
        "403":
          description: Requires ADMIN role
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_vndc_backend_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Unsuspend a user account (admin)
      tags:
      - Users
...
securityDefinitions:
  BearerAuth:
    description: 'JWT access token. Format: `Bearer <token>`'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...